+++ src/localization/localization.ts
@@ -1,4 +1,4 @@
-export type Localization = "EN-en" | "ES-es" | "FR-fr";
+export type Localization = "EN-en" | "ES-es" | "FR-fr" | "PT-pt";
 
 
 export type Components =